userInfoStore 분리
